+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 
   * Fixed linking error of FLANN by explicitly linking to lz4: [#1221](https://github.com/dartsim/dart/pull/1221)
 
-* Build system
+* Build System
 
   * Changed to use GNUInstallDirs for install paths: [#1241](https://github.com/dartsim/dart/pull/1241)
   * Fixed not failing for missing required dependencies: [#1250](https://github.com/dartsim/dart/pull/1250)
@@ -28,6 +28,10 @@
 * Dynamics
 
   * Fixed Skeleton::setState(): [#1245](https://github.com/dartsim/dart/pull/1245)
+
+* Python
+
+  * Added (experimental) Python binding: [#1237](https://github.com/dartsim/dart/pull/1237)
 
 #### Compilers Tested